refactor: comment 파일명에서 복수표현 제거
