+++ frontend/src/components/Comment/CommentList.tsx
@@ -1,4 +1,4 @@
-import { CommentRequest, CommentType } from '../../models/Comments';
+import { CommentRequest, CommentType } from '../../models/Comment';
 import Editor from '../Editor/Editor';
 import Comment from './Comment';
 import * as Styled from './CommentList.style';
