+++ frontend/src/mocks/handlers/index.ts
@@ -1,4 +1,4 @@
 import { popularStudyLogHandler } from './popularStudyLog';
-import { commentsHandler } from './comments';
+import { commentsHandler } from './comment';
 
 export const handlers = [...popularStudyLogHandler, ...commentsHandler];
